LIBIIIF-90. Attempt to get image info for fedora2 PIDs from Solr.
Query the fcrepo Solr core pcdm request handler to try and get the image dimensions for all the images that are part of a given UMDM PID. If this doesn't work, fall back to the old (and slow) way of asking the image server for the info.json for each image. Added a config key for the fedora2 handler, fcrepo_solr_url, that is the URL of the fedora4 core to use. Modified code that queries Solr to assume that the URLs provided in config are to the core, so the application just appends the request handler name.

https://issues.umd.edu/browse/LIBIIIF-90