LIBIIIF-99 Allow METS item label to be optional
Labels on METS fptr elements are optional, but the current pcdm-manifests code expects them to be there. The fix should be to apply a generic "Page N" label to any canvas where the METS data doesn't have a corresponding label, where N is that canvases 1-based index in the document order.

https://issues.umd.edu/browse/LIBIIIF-99